Move index path functions to utils
